+++ packages/common/src/project/versioned/v1_0_0/models.ts
@@ -79,10 +79,10 @@ export class BlockFilterImpl implements BlockFilter {
 
 export class ParentProjectModel implements ParentProject {
   @IsNumber()
-  @ValidateIf((obj, value) => value != null || obj.unitlBlock === null || obj.untilBlock === undefined)
+  @ValidateIf((obj, value) => value !== null || obj.unitlBlock === null || obj.untilBlock === undefined)
   block: number;
   @IsNumber()
-  @ValidateIf((obj, value) => value != null || obj.block === null || obj.block === undefined)
+  @ValidateIf((obj, value) => value !== null || obj.block === null || obj.block === undefined)
   untilBlock: number;
   @IsString()
   reference: string;